Martproject/User_Services/pyproject.toml
[tool.poetry]
name = "user-services"
version = "0.1.0"
description = ""
authors = ["Sajjad55555 <140962256+Sajjad55555@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.31.0"
fastapi = "^0.115.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"
sqlmodel = "^0.0.22"
psycopg2-binary = "^2.9.9"
aiokafka = "^0.11.0"
protobuf = "^5.28.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
